% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapreprocessing.R
\name{AggregateDataGP}
\alias{AggregateDataGP}
\title{Aggregate Data for Gaussian Process (GP) Algorithms}
\usage{
AggregateDataGP(PricesTested, PurchaseDecisions, TestX, sigma2_y, BatchSize)
}
\arguments{
\item{PricesTested}{A vector of prices (values between 0 and 1) that have been tested so far in the experiment.}

\item{PurchaseDecisions}{A vector of associated purchase decisions for the prices tested.}

\item{TestX}{A set of prices (values between 0 and 1) to test.}

\item{sigma2_y}{Observation noise variance (sigma_y^2), must be less than 0.25.}

\item{BatchSize}{Number of consumers tested before an algorithm can update.}
}
\value{
A data frame containing:
  \describe{
    \item{PricesTested}{The prices included in the dataset.}
    \item{PurchaseRates}{The mean purchase rate for each price.}
    \item{sigma_2y}{The adjusted observation noise variance for each price.}
  }
}
\description{
This function processes price testing data for Gaussian Process algorithms. It computes purchase rates
and adjusts observation noise variance for each price. Prices should be in the range [0, 1], and
observation noise variance (sigma2_y) must be less than 0.25.
}
\examples{
PricesTested <- c(0.1, 0.2, 0.2, 0.9)
PurchaseDecisions <- c(1, 0, 1, 0)
TestX <- seq(10)/10
sigma2_y <- rep(0.25, 10)
BatchSize <- 10
AggregateDataGP(PricesTested, PurchaseDecisions, TestX, sigma2_y, BatchSize)
}
