% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapreprocessing.R
\name{AggregateDataUCB}
\alias{AggregateDataUCB}
\title{GP Variants}
\usage{
AggregateDataUCB(PricesTested, PurchaseDecisions, TestX)
}
\arguments{
\item{PricesTested}{A vector of prices (values between 0 and 1) that have been tested so far in the experiment.}

\item{PurchaseDecisions}{A vector of associated purchase decisions for the prices tested.}

\item{TestX}{A set of prices (values between 0 and 1) to test.}
}
\value{
A list containing:
  \describe{
    \item{s_t}{The sum of purchase decisions for each price.}
    \item{n_t}{The number of observations for each price.}
  }
}
\description{
Collection of functions to aggregate data for various algorithms, including UCB, Thompson Sampling, and Gaussian Processes.
Aggregate Data for Upper Confidence Bound (UCB) Algorithm

This function processes price testing data for the UCB algorithm. It aggregates purchase decisions
by price, includes priors for prices that have not yet been tested, and returns summarized data.
Prices should be in the range [0, 1].
}
\examples{
PricesTested <- c(0.1, 0.2, 0.2, 0.9)
PurchaseDecisions <- c(1, 0, 1, 0)
TestX <- seq(10)/10
AggregateDataUCB(PricesTested, PurchaseDecisions, TestX)
}
