% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapostprocessing.R
\name{AppendExRewards}
\alias{AppendExRewards}
\title{Append Expected and Optimal Rewards to Simulation Data}
\usage{
AppendExRewards(
  SimData,
  Prices,
  NumIter,
  NumSimulations,
  SeasonLength,
  ExRewardsTV,
  SeasonOpt
)
}
\arguments{
\item{SimData}{A data frame containing simulation results.}

\item{Prices}{A vector of prices in the test set.}

\item{NumIter}{Integer. Number of iterations (consumers) in each simulation run.}

\item{NumSimulations}{Integer. Number of independent simulation runs.}

\item{SeasonLength}{Number of consumers per season; determines the frequency of seasonal demand shifts.}

\item{ExRewardsTV}{A matrix containing the expected rewards at every price in test set for every simulation (time varying).}

\item{SeasonOpt}{A numeric vector containing the repeated optimal values for each block in each simulation.}
}
\value{
A data frame identical to \code{SimData} but with additional columns for realized reward, 
action ID, iteration ID, season index, expected reward, and optimal reward.
}
\description{
This function augments a simulation data frame by computing and appending columns 
for realized rewards, expected rewards, and optimal rewards for each observation, 
using information on prices, expected reward matrices, and seasonal structure.
}
