% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{ArmsPlayedPlot}
\alias{ArmsPlayedPlot}
\title{Plot Histogram of Arms Played with Relative Reward Benchmarks}
\usage{
ArmsPlayedPlot(
  SelectedResults,
  OptVec,
  Title,
  PolicyNames,
  Colors,
  Prices,
  PriceGranularity,
  y_max = 100
)
}
\arguments{
\item{SelectedResults}{A list containing the relevant subset of the results for the plot.}

\item{OptVec}{A numeric vector indicating the expected reward of each price relative to the true optimal.}

\item{Title}{A character string to be used as the title of the plot.}

\item{PolicyNames}{A vector containing the policy names in the plot.}

\item{Colors}{A vector of colors for the plot.}

\item{Prices}{A vector of prices in the test set.}

\item{PriceGranularity}{A numeric value indicating the spacing between price levels.}

\item{y_max}{A numeric value specifying the maximum value of the y-axis (default is 100).}
}
\value{
A ggplot2 object representing the histogram of arms played, with overlaid benchmarks for relative expected rewards.
}
\description{
This function generates a bar plot showing the percentage of times each price (arm) was chosen 
by different bandit policies. A red benchmark line is overlaid to indicate how good each price is 
relative to the true optimal in terms of expected reward. The plot allows for visual comparison 
of policy behavior against the relative reward quality of each price.
}
