% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basisfunctions.R
\name{BasisFunction}
\alias{BasisFunction}
\title{Basis Function}
\usage{
BasisFunction(x, J)
}
\arguments{
\item{x}{Test point where the basis function is evaluated.}

\item{J}{Number of knots for the basis function.}
}
\value{
A vector of length J containing the basis function values at the test point.
}
\description{
Computes the values of a basis function at a test point, given a number of knots.
}
\details{
The function computes the basis function values using a piecewise linear approach with J knots. The calculation adjusts based on the location of the test point relative to the knot positions.
}
