% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedrewards.R
\name{ComputeExpectedRewardsFromCDF}
\alias{ComputeExpectedRewardsFromCDF}
\title{Expected Rewards from CDF}
\usage{
ComputeExpectedRewardsFromCDF(Prices, CDF_x, CDF_y)
}
\arguments{
\item{Prices}{A vector of prices in the test set.}

\item{CDF_x}{Vector of x-values (support) for the CDF.}

\item{CDF_y}{Vector of CDF values corresponding to \code{CDF_x}.}
}
\value{
A numeric vector of expected rewards, one for each price in \code{Prices}.
}
\description{
Computes the expected reward at each price using a given CDF. The expected reward is calculated as the price multiplied by the probability of purchase (1 - CDF at that price).
}
