% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{CovarianceFromKernel}
\alias{CovarianceFromKernel}
\title{Covariance Matrix from Kernel}
\usage{
CovarianceFromKernel(X1, X2, kernel, sigma_f, l)
}
\arguments{
\item{X1}{Matrix of m points (m x d).}

\item{X2}{Matrix of n points (n x d).}

\item{kernel}{Kernel function to compute covariance.}

\item{sigma_f}{Hyperparameter defining the vertical scale.}

\item{l}{Hyperparameter defining the horizontal scale.}
}
\value{
Covariance matrix of size m x n.
}
\description{
Computes the covariance matrix between two sets of points.
}
