% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{CumulativeRewardsPlot}
\alias{CumulativeRewardsPlot}
\title{Plot Cumulative Percentage of Optimal Rewards Over Time}
\usage{
CumulativeRewardsPlot(
  SelectedResults,
  PolicyNames,
  NumIter,
  Title,
  LineColors,
  LineTypes,
  RatioMaxTrue,
  legend_position = "none"
)
}
\arguments{
\item{SelectedResults}{A list containing the relevant subset of the results for the plot.}

\item{PolicyNames}{A vector containing the policy names in the plot.}

\item{NumIter}{Integer. Number of iterations (consumers) in each simulation run.}

\item{Title}{A character string to be used as the title of the plot.}

\item{LineColors}{A character vector of colors for each policy line.}

\item{LineTypes}{A vector of line types corresponding to each policy.}

\item{RatioMaxTrue}{Ratio between max reward possible from price set and true optimal (value between 0 and 1)}

\item{legend_position}{Character string specifying the position of the legend ("none", "bottom").}
}
\value{
A ggplot2 object representing the cumulative reward comparison plot.
}
\description{
This function generates a line plot showing the cumulative percentage of optimal rewards
achieved over time for different bandit policies. It visualizes performance over the course
of consumer interactions in a simulation, with optional customization for appearance.
}
