% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapostprocessing.R
\name{CumulativeRewardTV}
\alias{CumulativeRewardTV}
\title{Compute Cumulative Percentage of Optimal Rewards Over Time}
\usage{
CumulativeRewardTV(SimData, NumIter, NumSimulations)
}
\arguments{
\item{SimData}{A data frame containing simulation results.}

\item{NumIter}{Integer. Number of iterations (consumers) in each simulation run.}

\item{NumSimulations}{Integer. Number of independent simulation runs.}
}
\value{
A numeric vector of length \code{NumIter} giving the cumulative percentage of the optimal reward 
obtained at each time step.
}
\description{
This function calculates the cumulative percentage of the optimal reward achieved 
across all simulations over time, based on expected and optimal rewards in the simulation data.
}
