% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapostprocessing.R
\name{Histogram}
\alias{Histogram}
\title{Compute Normalized Histogram of Arms Played}
\usage{
Histogram(PricesTested, NumIter, Prices)
}
\arguments{
\item{PricesTested}{A vector of prices that were played during the experiment.}

\item{NumIter}{Integer. Number of iterations (consumers) in each simulation run.}

\item{Prices}{A vector of prices in the test set.}
}
\value{
A numeric vector of relative frequencies, one for each price in the test set.
}
\description{
This function computes a normalized histogram of the prices (arms) tested during an experiment,
comparing them against the full set of prices. It returns the relative frequency of each price being played.
}
