% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{Inverse}
\alias{Inverse}
\title{Inverse CDF}
\usage{
Inverse(u, CDF_x, CDF_y)
}
\arguments{
\item{u}{Point at which to find the inverse CDF.}

\item{CDF_x}{Vector of x-axis points of the CDF.}

\item{CDF_y}{Vector of the CDF values.}
}
\value{
The inverse CDF value at point u.
}
\description{
Computes the inverse of the CDF at a given point.
}
