% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{JointCovFromKernel}
\alias{JointCovFromKernel}
\title{Joint Covariance Matrix from Kernel}
\usage{
JointCovFromKernel(X, Index, kernel, sigma_f, l)
}
\arguments{
\item{X}{Vector of points.}

\item{Index}{Vector indicating the derivative order at each point (0 or 1).}

\item{kernel}{Kernel function to compute covariance.}

\item{sigma_f}{Hyperparameter defining the vertical scale.}

\item{l}{Hyperparameter defining the horizontal scale.}
}
\value{
Joint covariance matrix of size length(X) x length(X).
}
\description{
Computes the joint covariance matrix for a set of points and derivatives.
}
