% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{LeftDigitCDF}
\alias{LeftDigitCDF}
\title{Left-Digit CDF}
\usage{
LeftDigitCDF(CDFGranularity, DP, DG, ShapeA, ShapeB)
}
\arguments{
\item{CDFGranularity}{Step size for generating the x-axis points of the CDF.}

\item{DP}{Vector of discontinuity points.}

\item{DG}{Vector of discontinuity gaps.}

\item{ShapeA}{Beta distribution parameter.}

\item{ShapeB}{Beta distribution parameter.}
}
\value{
A vector of discrete CDF values for the left-digit case.
}
\description{
Creates a CDF with discontinuity gaps at specified points.
}
