% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{LeftDigitInverse}
\alias{LeftDigitInverse}
\title{Inverse Left-Digit CDF}
\usage{
LeftDigitInverse(u, CDFGranularity, LDCDF)
}
\arguments{
\item{u}{Point at which to find the inverse CDF.}

\item{CDFGranularity}{Step size for generating the x-axis points of the CDF.}

\item{LDCDF}{Vector of left-digit CDF values.}
}
\value{
The inverse CDF value at point u.
}
\description{
Computes the inverse of the left-digit CDF at a given point.
}
