% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpregression.R
\name{MakePosDefinitive}
\alias{MakePosDefinitive}
\title{Positive Definite Covariance Matrix}
\usage{
MakePosDefinitive(CovMatrix, zilon)
}
\arguments{
\item{CovMatrix}{Matrix to be corrected.}

\item{zilon}{Small value to replace negative eigenvalues.}
}
\value{
Positive definite covariance matrix.
}
\description{
Ensures a covariance matrix is positive definite by adjusting negative eigenvalues.
}
