% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabexperiment.R
\name{MultipleMABExperiments}
\alias{MultipleMABExperiments}
\title{Multi-Armed Bandit Experiments}
\usage{
MultipleMABExperiments(
  Seeds,
  DGPType,
  Policy,
  TestX,
  NumIter,
  BatchSize,
  NumKnots = NULL,
  Knots = NULL,
  BasisFunctions = NULL,
  HeteroNoise = FALSE,
  Reset = NULL,
  DistrType = NULL,
  Location = NULL,
  ShapeA = NULL,
  ShapeB = NULL,
  SeasonLength = NULL,
  Rho = NULL,
  CDFGranularity = NULL,
  OneCentScaled = NULL,
  DP = NULL,
  Zeta = NULL,
  CDF = NULL,
  Scale = NULL
)
}
\arguments{
\item{Seeds}{A vector of the length of the number of simulations allowing reproducibility.}

\item{DGPType}{The type of dgp ('static', 'timevarying', 'leftdigit', 'empirical').}

\item{Policy}{The policy function to be used in the experiment (e.g., UCB, TS, GPTS, GPUCB).}

\item{TestX}{A vector of prices (values between 0 and 1) to test.}

\item{NumIter}{Number of iterations (time steps) to run the experiment.}

\item{BatchSize}{Number of consumers tested before the policy updates its action scores.}

\item{NumKnots}{Number of knots for monotonic Gaussian Process variants (monotonic variants).}

\item{Knots}{A vector of knot locations for monotonic Gaussian Process regression (monotonic variants).}

\item{BasisFunctions}{A matrix of basis functions for monotonic Gaussian Process regression (monotonic variants).}

\item{HeteroNoise}{Logical; if `TRUE`, allows for heterogeneous noise modeling (optional, default is `FALSE`).}

\item{Reset}{Number of consumers before the experiment history is wiped (optional - time varying).}

\item{DistrType}{The distribution type ('beta', 'gumbel', 'frechet', 'gev') (optional).}

\item{Location}{Distribution parameter (optional).}

\item{ShapeA}{Distribution parameter (optional).}

\item{ShapeB}{Distribution parameter (optional).}

\item{SeasonLength}{Length of each season. (timevarying)}

\item{Rho}{Maximum seasonal shock (timevarying).}

\item{CDFGranularity}{Step size for generating the x-axis points of the CDF (leftdigit).}

\item{OneCentScaled}{Value of one cent when scaled to 0-1 range (leftdigit).}

\item{DP}{Vector of discontinuity points (leftdigit).}

\item{Zeta}{Scaling factor for the gap aka jump parameter (leftdigit).}

\item{CDF}{A data frame representing the empirical CDF. The first column 
contains WTP, and the second column contains cumulative probabilities (empirical).}

\item{Scale}{Numeric; the scaling factor applied to the WTP data to ensure 
prices are normalized (e.g., between 0 and 1) (empirical).}

\item{DistrScale}{Distribution parameter (optional).}
}
\value{
A data frame containing three columns:
  \describe{
    \item{PricesTested}{A vector of prices tested over the experiment.}
    \item{PurchaseDecisions}{A vector of binary purchase decisions corresponding to each price tested.}
    \item{SimulationID}{A vector containing the Simulation ID corresponding to the data}
  }
}
\description{
Allows for multiple MAB experiment simulations to be run
}
