% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneityfunctions.R
\name{NoiseSample}
\alias{NoiseSample}
\title{Noise Sampling for Heteroscedastic Gaussian Processes}
\usage{
NoiseSample(PricesTested, PurchaseDecisions, TestX, BatchSize)
}
\arguments{
\item{PricesTested}{A vector of prices (values between 0 and 1) that have been tested so far in the experiment.}

\item{PurchaseDecisions}{A vector of associated purchase decisions for the prices tested.}

\item{TestX}{A set of prices (values between 0 and 1) to test.}

\item{BatchSize}{Number of consumers tested before an algorithm can update.}
}
\value{
A vector containing the sampled heteroscedastic noise variances (\eqn{\sigma^2_y}).
}
\description{
This function generates a sample of heteroscedastic noise variance (\eqn{\sigma^2_y}) for a Gaussian Process.
The function processes price testing data and uses posterior predictions to generate a random draw.
Prices should be in the range [0, 1], and observation noise variance (\eqn{\sigma^2_y}) must be less than 0.25.
}
\examples{
PricesTested <- c(0.1, 0.2, 0.2, 0.9)
PurchaseDecisions <- c(1, 0, 1, 0)
TestX <- seq(10)/10
sigma2_y <- rep(0.25, 10)
BatchSize <- 10
NoiseSample(PricesTested, PurchaseDecisions, TestX, BatchSize)
}
