% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{ObtainDiscontinuityGaps}
\alias{ObtainDiscontinuityGaps}
\title{Discontinuity Gap Calculation}
\usage{
ObtainDiscontinuityGaps(OneCentScaled, ShapeA, ShapeB, DP, JumpParameter)
}
\arguments{
\item{OneCentScaled}{Value of one cent when scaled to 0-1 range.}

\item{ShapeA}{Beta distribution parameter.}

\item{ShapeB}{Beta distribution parameter.}

\item{DP}{Vector of discontinuity points.}

\item{JumpParameter}{Scaling factor for the gap.}
}
\value{
A vector of discontinuity gaps.
}
\description{
Calculates discontinuity gaps in demand at specified price points.
}
