% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpregression.R
\name{OptimalHyperparameters}
\alias{OptimalHyperparameters}
\title{Optimal Hyperparameters}
\usage{
OptimalHyperparameters(TrainX, TrainY, sigma2_y)
}
\arguments{
\item{TrainX}{Matrix of training points (m x d).}

\item{TrainY}{Vector of training targets (m x 1).}

\item{sigma2_y}{Observation noise variance (sigma_y^2).}
}
\value{
Vector of optimized hyperparameters (sigma_f and l).
}
\description{
Optimizes hyperparameters for Gaussian Process regression using NLopt.
}
