% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{PolicyAsString}
\alias{PolicyAsString}
\title{Extract the Name of a Policy Function}
\usage{
PolicyAsString(Policy)
}
\arguments{
\item{Policy}{A function representing the policy whose name is to be extracted.}
}
\value{
A character vector containing the name of the policy function.
}
\description{
This function takes a policy function as input and returns its name as a string.
}
\examples{
MyPolicy <- UCB
PolicyAsString(MyPolicy)
}
