% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpregression.R
\name{PosteriorPrediction}
\alias{PosteriorPrediction}
\title{Posterior Prediction (Joint GP with Derivatives)}
\usage{
PosteriorPrediction(TrainX, TrainY, TestX, TestD, sigma_f, l, sigma2_y)
}
\arguments{
\item{TrainX}{Vector of training points.}

\item{TrainY}{Vector of training targets.}

\item{TestX}{Vector of test points.}

\item{TestD}{Vector of test points for derivative prediction.}

\item{sigma_f}{Hyperparameter for kernel vertical scale.}

\item{l}{Hyperparameter for kernel horizontal scale.}

\item{sigma2_y}{Observation noise variance (sigma_y^2).}
}
\value{
List containing posterior mean vector and covariance matrix.
}
\description{
Computes the posterior mean and covariance for Gaussian Process regression with derivatives.
}
