% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{RBFKernel_01}
\alias{RBFKernel_01}
\title{RBF Kernel (Point to Derivative)}
\usage{
RBFKernel_01(x_i, x_j, sigma_f, l)
}
\arguments{
\item{x_i}{A point with d dimensions.}

\item{x_j}{A point (derivative) with d dimensions.}

\item{sigma_f}{Hyperparameter defining the vertical scale.}

\item{l}{Hyperparameter defining the horizontal scale.}
}
\value{
Kernel value between the point and derivative.
}
\description{
Calculates the RBF kernel between a point and derivative.
}
