% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{RBFKernel_11}
\alias{RBFKernel_11}
\title{RBF Kernel (Derivative to Derivative)}
\usage{
RBFKernel_11(x_i, x_j, sigma_f, l)
}
\arguments{
\item{x_i}{A point (derivative) with d dimensions.}

\item{x_j}{A point (derivative) with d dimensions.}

\item{sigma_f}{Hyperparameter defining the vertical scale.}

\item{l}{Hyperparameter defining the horizontal scale.}
}
\value{
Kernel value between the two derivatives.
}
\description{
Calculates the RBF kernel between two derivatives.
}
