% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{TS}
\alias{TS}
\title{Thompson Sampling (TS) Policy}
\usage{
TS(PricesTested, PurchaseDecisions, TestX)
}
\arguments{
\item{PricesTested}{A vector of prices (values between 0 and 1) that have been tested so far in the experiment.}

\item{PurchaseDecisions}{A vector of associated purchase decisions for the prices tested.}

\item{TestX}{A set of prices (values between 0 and 1) to test.}
}
\value{
A vector of action scores for the prices in `TestX`.
}
\description{
This function implements the Thompson Sampling policy for pricing experiments. It calculates
action scores by drawing from a Beta distribution based on aggregated data.
}
\examples{
PricesTested <- c(0.1, 0.2, 0.2, 0.9)
PurchaseDecisions <- c(1, 0, 1, 0)
TestX <- seq(10)/10
TS(PricesTested, PurchaseDecisions, TestX)
}
