% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{UCB}
\alias{UCB}
\title{Bandit Policies for Pricing Experiments}
\usage{
UCB(PricesTested, PurchaseDecisions, TestX)
}
\arguments{
\item{PricesTested}{A vector of prices (values between 0 and 1) that have been tested so far in the experiment.}

\item{PurchaseDecisions}{A vector of associated purchase decisions for the prices tested.}

\item{TestX}{A set of prices (values between 0 and 1) to test.}
}
\value{
A vector of action scores for the prices in `TestX`.
}
\description{
Collection of functions implementing various policies for pricing experiments, including Upper Confidence Bound (UCB), Thompson Sampling (TS), Gaussian Process-based policies, and their monotonic extensions.
Upper Confidence Bound (UCB) Policy

This function implements the UCB policy for pricing experiments. It calculates action scores
based on aggregated data and selects prices to test using an upper confidence bound.
}
\examples{
PricesTested <- c(0.1, 0.2, 0.2, 0.9)
PurchaseDecisions <- c(1, 0, 1, 0)
TestX <- seq(10)/10
UCB(PricesTested, PurchaseDecisions, TestX)
}
