% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{ValuationSample}
\alias{ValuationSample}
\title{Wrapper for Valuation Sample Functions}
\usage{
ValuationSample(
  DGPType,
  NumSamples,
  DistrType = NULL,
  Location = NULL,
  DistrScale = NULL,
  ShapeA = NULL,
  ShapeB = NULL,
  SeasonLength = NULL,
  Rho = NULL,
  CDFGranularity = NULL,
  OneCentScaled = NULL,
  DP = NULL,
  Zeta = NULL,
  CDF = NULL,
  Scale = NULL
)
}
\arguments{
\item{DGPType}{The type of dgp ('static', 'timevarying', 'leftdigit', 'empirical').}

\item{NumSamples}{Number of valuations to sample.}

\item{DistrType}{The distribution type ('beta', 'gumbel', 'frechet', 'gev') (optional).}

\item{Location}{Distribution parameter (optional).}

\item{DistrScale}{Distribution parameter (optional).}

\item{ShapeA}{Distribution parameter (optional).}

\item{ShapeB}{Distribution parameter (optional).}

\item{SeasonLength}{Length of each season. (timevarying)}

\item{Rho}{Maximum seasonal shock (timevarying).}

\item{CDFGranularity}{Step size for generating the x-axis points of the CDF (leftdigit).}

\item{OneCentScaled}{Value of one cent when scaled to 0-1 range (leftdigit).}

\item{DP}{Vector of discontinuity points (leftdigit).}

\item{Zeta}{Scaling factor for the gap aka jump parameter (leftdigit).}

\item{CDF}{A data frame representing the empirical CDF. The first column 
contains WTP, and the second column contains cumulative probabilities (empirical).}

\item{Scale}{Numeric; the scaling factor applied to the WTP data to ensure 
prices are normalized (e.g., between 0 and 1) (empirical).}
}
\value{
A numeric vector of sampled valuations.
}
\description{
Generates samples from one of the different types of distributions
}
