% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{ValuationSampleLD}
\alias{ValuationSampleLD}
\title{Left-Digit Valuation Sample}
\usage{
ValuationSampleLD(NumSamples, CDFGranularity, LDCDF)
}
\arguments{
\item{NumSamples}{Number of valuations to sample.}

\item{CDFGranularity}{Step size for generating the x-axis points of the CDF.}

\item{LDCDF}{Vector of left-digit CDF values.}
}
\value{
A vector of sampled valuations.
}
\description{
Generates samples from a left-digit adjusted distribution of valuations.
}
