% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{GPTS}
\alias{GPTS}
\title{Gaussian Process Thompson Sampling (GPTS) Policy}
\usage{
GPTS(PricesTested, PurchaseDecisions, TestX, sigma2_y, BatchSize)
}
\arguments{
\item{PricesTested}{A vector of prices (values between 0 and 1) that have been tested so far in the experiment.}

\item{PurchaseDecisions}{A vector of associated purchase decisions for the prices tested.}

\item{TestX}{A set of prices (values between 0 and 1) to test.}

\item{sigma2_y}{Observation noise variance (must be less than 0.25).}

\item{BatchSize}{Number of consumers tested before an algorithm can update.}
}
\value{
A vector of action scores for the prices in `TestX`.
}
\description{
This function implements the GPTS policy for pricing experiments. It uses Gaussian Process regression
to generate posterior predictions and action scores.
}
\examples{
PricesTested <- c(0.1, 0.2, 0.2, 0.9)
PurchaseDecisions <- c(1, 0, 1, 0)
TestX <- seq(10)/10
sigma2_y <- rep(0.25, 10)
BatchSize <- 10
GPTS(PricesTested, PurchaseDecisions, TestX, sigma2_y, BatchSize)
}
