% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policies.R
\name{GPUCB_Mono}
\alias{GPUCB_Mono}
\title{Gaussian Process Upper Confidence Bound Monotonic (GPUCB_Mono) Policy}
\usage{
GPUCB_Mono(
  PricesTested,
  PurchaseDecisions,
  TestX,
  Knots,
  sigma2_y,
  BatchSize,
  BasisFunctions,
  LB1,
  LB2,
  UB
)
}
\arguments{
\item{PricesTested}{A vector of prices (values between 0 and 1) that have been tested so far in the experiment.}

\item{PurchaseDecisions}{A vector of associated purchase decisions for the prices tested.}

\item{TestX}{A set of prices (values between 0 and 1) to test.}

\item{Knots}{Knots for the basis functions used in the regression.}

\item{sigma2_y}{Observation noise variance (must be less than 0.25).}

\item{BatchSize}{Number of consumers tested before an algorithm can update.}

\item{BasisFunctions}{Basis functions for the monotonic Gaussian Process regression.}

\item{LB1}{Lower bound for the test prices.}

\item{LB2}{Lower bound for the knots.}

\item{UB}{Upper bound for both test prices and knots.}
}
\value{
A vector of action scores for the prices in `TestX`.
}
\description{
This function implements the monotonic Gaussian Process Upper Confidence Bound (GPUCB_Mono) policy for pricing experiments.
It uses Gaussian Process regression to calculate action scores based on posterior predictions and confidence intervals 
while ensuring monotonicity.
}
