% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabexperiment.R
\name{MABExperiment}
\alias{MABExperiment}
\title{Multi-Armed Bandit Experiment Framework}
\usage{
MABExperiment(
  Valuations,
  Policy,
  TestX,
  NumIter,
  BatchSize,
  NumKnots = NULL,
  Knots = NULL,
  BasisFunctions = NULL,
  HeteroNoise = FALSE,
  Reset = NULL
)
}
\arguments{
\item{Valuations}{A vector of consumer valuations for the product, used to simulate purchase decisions.}

\item{Policy}{The policy function to be used in the experiment (e.g., UCB, TS, GPTS, GPUCB).}

\item{TestX}{A vector of prices (values between 0 and 1) to test.}

\item{NumIter}{Number of iterations (time steps) to run the experiment.}

\item{BatchSize}{Number of consumers tested before the policy updates its action scores.}

\item{NumKnots}{Number of knots for monotonic Gaussian Process variants (optional, default is `NULL`).}

\item{Knots}{A vector of knot locations for monotonic Gaussian Process regression (optional, default is `NULL`).}

\item{BasisFunctions}{A matrix of basis functions for monotonic Gaussian Process regression (optional, default is `NULL`).}

\item{HeteroNoise}{Logical; if `TRUE`, allows for heterogeneous noise modeling (optional, default is `FALSE`).}

\item{Reset}{Number of consumers before the experiment history is wiped (optional, default is `NULL`).}
}
\value{
A data frame containing two columns:
  \describe{
    \item{PricesTested}{A vector of prices tested over the experiment.}
    \item{PurchaseDecisions}{A vector of binary purchase decisions corresponding to each price tested.}
  }
}
\description{
This file contains functions for running multi-armed bandit (MAB) experiments, supporting various policies, Gaussian Process-based methods, and monotonic variants. It allows simulation of consumer purchase decisions under different pricing strategies.
Run a Multi-Armed Bandit (MAB) Experiment

This function simulates a Multi-Armed Bandit experiment using a specified policy and test settings.
It supports both Gaussian Process-based and monotonic variants, with optional heterogeneous noise modeling.
}
\examples{
Valuations <- runif(100, min = 0.1, max = 0.9)
TestX <- seq(10)/10
NumIter <- 100
BatchSize <- 10
MABExperiment(Valuations, UCB, TestX, NumIter, BatchSize)
}
