% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PricingBandits-package.R
\docType{package}
\name{PricingBandits-package}
\alias{PricingBandits}
\alias{PricingBandits-package}
\title{PricingBandits: Multi-Armed Bandit Approaches to Pricing Experiments}
\description{
Implements multi-armed bandit approaches for pricing experiments with an unknown demand curve, as developed in Weaver, Kumar, and Jain, "Nonparametric Pricing Bandits Leveraging Informational Externalities to Learn the Demand Curve" (Marketing Science). Includes Upper Confidence Bound (UCB) and Thompson Sampling (TS) baselines, Gaussian process variants ('GP-UCB', 'GP-TS'), monotonic Gaussian process variants that constrain demand to be weakly decreasing in price, and heterogeneous-noise extensions. The willingness-to-pay distribution is fully user-specified via a vector of consumer valuations, so any demand environment can be simulated or replayed.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ian-weaver/PricingBandits}
  \item Report bugs at \url{https://github.com/ian-weaver/PricingBandits/issues}
}

}
\author{
\strong{Maintainer}: Ian N. Weaver \email{i.weaver@nus.edu.sg}

Authors:
\itemize{
  \item Ian N. Weaver \email{i.weaver@nus.edu.sg}
  \item Vineet Kumar
  \item Lalit Jain
}

}
\keyword{internal}
