% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{ValuationSampleLD}
\alias{ValuationSampleLD}
\title{Left-Digit Valuation Sample}
\usage{
ValuationSampleLD(
  NumSamples,
  CDFGranularity,
  OneCentScaled,
  ShapeA,
  ShapeB,
  DP,
  Zeta
)
}
\arguments{
\item{NumSamples}{Number of valuations to sample.}

\item{CDFGranularity}{Step size for generating the x-axis points of the CDF.}

\item{OneCentScaled}{Value of one cent when scaled to 0-1 range.}

\item{ShapeA}{Beta distribution parameter.}

\item{ShapeB}{Beta distribution parameter.}

\item{DP}{Vector of discontinuity points.}

\item{Zeta}{Scaling factor for the gap (jump parameter).}
}
\value{
A vector of sampled valuations.
}
\description{
Generates samples from a left-digit adjusted distribution of valuations.
}
